+++ docs/markdown/changelog.md
@@ -5,9 +5,27 @@
 ### New
 
 
+* Instal.sh script. [pcotte]
+
+
+* Added doc. [pcotte]
+
+
+* Added doc. [pcotte]
+
+
 * First commit. [pcotte]
 
+### Changes
+
+
+* Clearer myisinstance. [pcotte]
+
 ### Fix
+
+
+* Improved myisinstance by including the possibility to specify a list,
+set or tuple of types. [pcotte]
 
 
 * Added more versions to pytests. [pcotte]